requirements.txt
streamlit
playwright
openai
langchain
psycopg2-binary
python-dotenv
supabase
pandas
numpy
aiohttp
asyncio
python-dateutil
pytz
requests
urllib3
beautifulsoup4
black
pylint
pytest
pytest-asyncio